+++ app/controllers/facilities_management/rm6378/admin/frameworks_controller.rb
@@ -2,7 +2,7 @@ module FacilitiesManagement
   module RM6378
     module Admin
       class FrameworksController < FacilitiesManagement::Admin::FrameworkController
-        include ::Admin::FrameworksController
+        include ::Admin::FrameworkActions
       end
     end
   end