+++ app/controllers/facilities_management/rm6378/admin/reports_controller.rb
@@ -2,7 +2,7 @@ module FacilitiesManagement
   module RM6378
     module Admin
       class ReportsController < FacilitiesManagement::Admin::FrameworkController
-        include ::Admin::ReportsController
+        include ::Admin::ReportActions
       end
     end
   end