+++ app/controllers/facilities_management/rm6378/admin/suppliers_controller.rb
@@ -2,11 +2,8 @@ module FacilitiesManagement
   module RM6378
     module Admin
       class SuppliersController < FacilitiesManagement::Admin::FrameworkController
-        include ::Admin::SuppliersController
-
-        SECTION_TO_PARAMS = {
-          basic_supplier_information: %i[name duns_number sme],
-        }.freeze
+        include ::Admin::SupplierActions
+        include SectionsConcern
       end
     end
   end