+++ spec/models/framework_spec.rb
@@ -10,6 +10,7 @@ RSpec.describe Framework do
     it { is_expected.to have_many(:procurements) }
     it { is_expected.to have_many(:uploads) }
     it { is_expected.to have_many(:admin_uploads) }
+    it { is_expected.to have_many(:change_logs) }
   end
 
   describe '.find_scope_by_service_name' do