+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add the inital admin pages for Facilities Management RM6378 ([PR 5318](https://github.com/Crown-Commercial-Service/crown-marketplace/pull/5318))
 - Add the managing of supplier and lot data for Facilities Management RM6378 ([PR 5318](https://github.com/Crown-Commercial-Service/crown-marketplace/pull/5318))
+- Add change log records for admin updates to supplier data ([PR 5336](https://github.com/Crown-Commercial-Service/crown-marketplace/pull/5336))
+- Add change logs download in the admin section ([PR 5336](https://github.com/Crown-Commercial-Service/crown-marketplace/pull/5336))
 
 ### Changed
 
@@ -18,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove the "use gca" flags ([PR 5276](https://github.com/Crown-Commercial-Service/crown-marketplace/pull/5276))
 - Update how we do releases to preview and production to target a specific tag ([PR 5308](https://github.com/Crown-Commercial-Service/crown-marketplace/pull/5308))
 - Allow framework status to be updated at the framework level rather than the service level ([PR 5318](https://github.com/Crown-Commercial-Service/crown-marketplace/pull/5318))
+- Refactored the admin name spacing for Legal Services and Supply Teachers ([PR 5336](https://github.com/Crown-Commercial-Service/crown-marketplace/pull/5336))
 
 ### Fixed